+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.2.8] - 2026-09-08
 
 ### Added
 
@@ -162,6 +162,7 @@ Thanks to [@elianiva](https://github.com/elianiva) for identifying the storage i
 
 - Initial public release.
 
+[0.2.8]: https://github.com/yusuf8834/bb-sidebar/compare/v0.2.7...v0.2.8
 [0.2.7]: https://github.com/yusuf8834/bb-sidebar/compare/v0.2.6...v0.2.7
 [0.2.6]: https://github.com/yusuf8834/bb-sidebar/compare/v0.2.5...v0.2.6
 [0.2.5]: https://github.com/yusuf8834/bb-sidebar/compare/v0.2.4...v0.2.5